task_manager_be/requirements.txt
asgiref 
certifi 
charset-normalizer==3.1.0
click
colorama
Django
django-cors-headers==3.14.0
django-environ==0.10.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
environ==1.0
idna==3.4
psycopg2
psycopg2-binary==2.9.6
python-dotenv
pytz==2023.3
requests==2.28.2
sqlparse
tzdata
urllib3==1.26.15
wincertstore==0.2
gunicorn==20.1.0
python-dateutil==2.8.2
passlib==1.7.4
